requirements.txt
tensorflow==2.15.0
streamlit==1.29.0
numpy
pillow
opencv-python-headless